% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesX.R
\name{bayesX}
\alias{bayesX}
\title{execute a command on bayesX CLI}
\usage{
bayesX(prg_path, ...)
}
\arguments{
\item{prg_path}{program (in 'bayesX' language) to execute on bayesX}
}
\value{
an object of type 'bayesXResult' which contains the output generated
by bayesX CLI
}
\description{
execute a command on bayesX CLI
}

