% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesXOutput.R
\name{bayesXOutput}
\alias{bayesXOutput}
\alias{bayesXOutput.character}
\title{'bayesXOutput' + 'effects' constructor for 'output/' files generated by bayesX CLI}
\usage{
bayesXOutput(bayesXResult, ...)

\method{bayesXOutput}{character}(path, ...)
}
\arguments{
\item{bayesXResult}{is the object generated by \code{\link{bayesX}}}
}
\description{
class which contains information about 'output/' files generated by
\code{\link{bayesX}}
}
\note{
since user only passes the output directory, we do not know where
data sample is stored
}

