% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesX.R
\name{extract}
\alias{extract}
\title{extract information of bayesXResult}
\usage{
extract(bayesXResult, pattern, ...)
}
\arguments{
\item{pattern}{see in \code{\link{grep}}}

\item{...}{further arguments passed to \code{\link{grep}}}
}
\value{
a \code{list} with successfull match (\code{character}) and where
\code{numeric} match in bayesXResult happend
}
\description{
extract information of bayesXResult
}
\examples{
\dontrun{extract(res, "Acceptance rate")
library(BayesXShinyApp)
res <- bayesX("./inst/template.prg")
extract(res, "Acceptance rate")
extract(res, "Variance")
extract(res, "Minimum")
extract(res, "Maximum")
ex <- extract(res, "mean")
rbind(res[ex$where-1], ex$match, res[ex$where+1])}

}

