% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{output_path}
\alias{output_path}
\title{get output path generated by bayesX CLI}
\usage{
output_path(bayesXResult, ...)
}
\arguments{
\item{bayesXResult}{is the object generated by \code{\link{bayesX}}}
}
\value{
\code{character} vector which can be interpreted as path
}
\description{
get output path generated by bayesX CLI
}
\examples{
\dontrun{output_path(res)
list.files(output_path(res),full.names = TRUE)}
}

