% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameters}
\alias{parameters}
\title{recieve distributional parameters (e.g \strong{\eqn{\eta}}'s)}
\usage{
parameters(bayesXOutput, ...)
}
\description{
returns sample of distributional parameters (e.g \strong{\eqn{\eta}})
}

