% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{runs ShinyApp}
\usage{
run_app(display.mode = "normal", which = "shiny-app", ...)
}
\arguments{
\item{display.mode}{see \code{\link{runApp}}}

\item{...}{further arguments passed to \code{\link{runApp}}}
}
\description{
runs ShinyApp
}
\note{
arguments send to \code{\link{runApp}}
}
\examples{
\dontrun{run_app()}
}

