% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesXOutput.R
\name{[.bayesXOutput}
\alias{[.bayesXOutput}
\title{subscript operator for 'bayesXOutput' objects}
\usage{
\method{[}{bayesXOutput}(bayesXOutput, i, drop = TRUE, ...)
}
\arguments{
\item{bayesXOutput}{is the object generated by \code{\link{bayesXOutput}}}
}
\description{
extracts meta information out of 'output/*.r' file generated by
\code{\link{bayesX}}
}

