% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{supported_distributions}
\alias{supported_distributions}
\title{#' @export
distribution <- function(name, density, params, link){
  
  # create an instance of 'distribution' object
  not_fun <- !is.function(density) || !is.function(link)
  if( not_fun )
    stop("'denisty' and 'link' have to be of type 'function'")
  
  params_have_no_names <- is.null(attr(params, "names"))
  if( params_have_no_names )
    stop("all elements in 'params' need a name")
  
  structure(list(density = density, parameter = params, link = link), names = name)
}}
\usage{
supported_distributions()
}
\description{
#' @export
add <- function(obj) UseMethod("add")
add.distribution <- function(obj){
  .distributions <<- append(.distributions, obj, after = 0)
}
}

