% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{bayesXOutput}
\alias{bayesXOutput}
\title{'bayesXOutput' constructor for 'output/' files generated by bayesX CLI}
\usage{
bayesXOutput(bayesXResult, ...)
}
\arguments{
\item{bayesXResult}{is the object generated by \code{\link{bayesX}}}
}
\description{
class which contains information about 'output/' files generated by
\code{\link{bayesX}}
}

