% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesX.R
\name{data_}
\alias{data_}
\title{get data used in bayesX supplied program}
\usage{
data_(bayesXResult, ...)
}
\arguments{
\item{bayesXResult}{is the object generated by \code{\link{bayesX}}}
}
\description{
get data used in bayesX supplied program
}

