% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key_values}
\alias{get_key_values}
\title{extracts key value pairs from 'output' directory generated by bayesX CLI}
\usage{
get_key_values(line, data_env, ...)
}
\description{
extracts key value pairs from 'output' directory generated by bayesX CLI
}
\note{
we do not export this method, is an internal method
}

