% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{[.parameters}
\alias{[.parameters}
\title{subscript operator for 'parameters' objects}
\usage{
\method{[}{parameters}(parameters, ...)
}
\description{
subscript operator for 'parameters' objects
}
\note{
one can use \code{\link{all.equal}} to extract rows. Row's represent
the underlying grid of covariates. Covariates cannot be compared with
\code{==} operator, since they are \code{numeric}. Therefore one prefers
\code{\link{all.equal}} for 'near equality'.
}

